requirements.txt
streamlit==1.35.0
pdfplumber==0.11.0
fpdf2==2.7.9
pandas==2.2.2
